Last CuPy Test
Achieved a speedup of up to 28.86 for the GPU over the CPU with a matrix dimension of N = 30000 by adding a GPU warmup function to ensure that the CUDA context is initialized before the performance comparison. 

Next week I will try with C, using the NVidia CUDA Visual Studio Integration. Would be interesting to try cuBLAS. There are some integrations with Fortran already.